backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
httpx==0.28.0
anthropic==0.40.0
jinja2==3.1.4
pypdf==5.1.0
python-dotenv==1.0.1
redis==5.2.0
pydantic==2.10.3
pydantic-settings==2.6.1
slowapi==0.1.9
cryptography==44.0.0
